Recreate `AssetChecksDefinition`s after setting prefix (#25271)
## Summary & Motivation

This makes sure that the `load_asset_checks_from_*` functions return
`[AssetChecksDefinition]` by replacing the typecast with the creation of
an actual `AssetChecksDefinition`.

Alternatively, this issue could be addressed by adding a specialized
`with_attributes()` method to `AssetChecksDefinition` to make it return
the correct type. I'd be happy to try this path as well! But it feels
like this change would be sufficient and minimally impactful across the
codebase.

Closes https://github.com/dagster-io/dagster/issues/22064

## How I Tested These Changes

Added an `isinstance` check to the `load_asset_checks_from_*` tests