Make existing conditions work with Asset Checks (#25170)
## Summary & Motivation

This updates our suite of built-in automation conditions to work natively with asset checks (previously, they only worked with assets).

This means you can do things like use the `on_cron` or `eager` conditions with checks.

This requires a unified set of statuses that apply to both assets and asset checks, as each has different ways of describing their status.

## How I Tested These Changes

## Changelog

- The `AutomationCondition.eager()`, `AutomationCondition.missing()`, and `AutomationCondition.on_cron` conditions are now compatible with asset checks.