[dagster-airlift] Make proxy operator pluggable (#25205)
## Summary & Motivation
It's highly likely that the default magic we perform to "swizzle in" the
dagster operator will fail for some cases. We need to have fallbacks for
edge cases and weirdness with forked versions of airflow, weird
parameters, etc.

This introduces a few different pluggability points:
- Top-level callback to the `proxying_to_dagster` fxn which allows for
complete control of the creation of the new operator.
- Make the default `build_dagster_task` pluggable by allowing for
parameter level overrides, and the class to be instantiated to be
overridden. This should account for the common cases.

## How I Tested These Changes
Added a new kitchen sink test which invokes both layers of pluggability.
## Changelog
NOCHANGELOG