Create backfill_in_progress AutomationCondition (#25190)
## Summary & Motivation

As title. This converts the existing `in_progress` AutomationCondition to capture both in progress runs and in progress backfills. It keeps these as separate conditions to maximize flexibility (and also increases explainability in the UI, as you can tell if an asset is part of a backfill but not an active run, or vice-versa).

## How I Tested These Changes

## Changelog

`AutomationCondition.in_progress()` now will be true if an asset partition is part of an in-progress backfill that has not yet executed it. The prior behavior, which only considered runs, is encapsulated in `AutomationCondition.execution_in_progress()`.