[dagster-embedded-elt] clean up extraction of dlt metadata (#25307)
## Summary & Motivation
[Previous PR](https://github.com/dagster-io/dagster/pull/25266) needed
some cleaning up as we don't actually need to call as many methods and
functions as I thought at first. I also found a counter-example to the
solution when using double underscores in the table name which would
still fail to extract metadata. So the code should be cleaner and cover
more table identifier cases.

I also spotted a `dlt_source.discover_schema()` that was not supposed to
be there.

## How I Tested These Changes
`make ruff`
`pythom -m pytest`

Make the test dlt source use `Repo__Issues` as `table_name` to make sure
we can extract metadata when the table name contains double underscores.