[dagster-looker] Add support for refinements, extensions (#25321)
## Summary & Motivation

Adds support for
[refinements](https://cloud.google.com/looker/docs/lookml-refinements)
(which let you modify an explore or view after creating it) and
[extends](https://cloud.google.com/looker/docs/reusing-code-with-extends)
(which let you extend views or explores to create new ones):

```lkml
view: base {
  sql_table_name: `prod.base_data` ;;
  description: "This is the base view"

  dimension: email {
    type: string
    sql: ${TABLE}.email ;;
  }

  dimension: name {
    hidden: yes
    type: string
    sql: ${TABLE}.name ;;
  }
}

# Modify base to change table name, show name dimension
view: +base {
  sql_table_name: `prod.new_base_data` ;;

  dimension: name {
    hidden: no
    type: string
    sql: ${TABLE}.name ;;
  }
}
```

We now merge any changes from refinements into the config of the explore
or view before converting it to a spec. For explores or views which
extend other models, we topo sort all explores/views based on extend
dependencies and merge them accordingly.

## How I Tested These Changes

New unit tests.

## Changelog

> [dagster-looker] Added support for views and explores that rely on
refinements or extends.