Fix bug when switching from non-user-code to user-code (#25381)
## Summary & Motivation

As title. Previous codepaths assumed that for a given instigator, the `handled_by_asset_daemon` property would never change, which is not the case. To avoid this issue, we ignore the instigator state in these codepaths, and instead use the current value based off of the workspace snapshot.

## How I Tested These Changes

Added test which failed before these changes, and now passes.

## Changelog

Fixed an issue which would cause `AutomationConditionSensorDefinitions` to not be evaluated if the `use_user_code_server` value was toggled after the initial evaluation.