[ui] Update OperationQueryStringApolloLink to accept URI from context (#25401)
## Summary & Motivation

I'm making an internal change that adds a link that will modify the
GraphQL URI depending on the operation, so the OSS link that adds the
`op` query param needs to be updated to accommodate that URI value (if
any).

- Use the `uri` value from context or fall back to the `basePath`-based
value.
- Create a utility function that appends query params to the URI. I'll
use this to append an additional param to visually flag requests where
the URI has been changed.

## How I Tested These Changes

With the internal change in place, perform queries in Plus app. Verify
that queries are performed successfully and have the `op` tag appended.

For queries that have the new URI, verify that these are also
successful, and that the `op` tag and new query param are applied.