Propagate automation condition information through the UI for user code conditions (#25211)
## Summary & Motivation

It is not possible to serialize the full automation condition object if it is user-defined. In these cases, we only persist the snapshot.

Updates the graphql code to use this snapshot instead of the raw condition, and adds e2e test ensuring we can execute / get the correct label for these objects.

## How I Tested These Changes

## Changelog

NOCHANGELOG