Add test demonstrating that we do not hit the RESOURCE_EXHAUSTED error above a given threshold (#25214)
## Summary & Motivation

As title. This sets a baseline for feeling "safe" with this many assets.

Some upgrades were needed to the test suite to make it more realistic such that we could trip the error.

Future changes will allow us to greatly increase, or entirely remove, this limit.

## How I Tested These Changes

Confirmed that if 1000 assets were supplied, we would fail this test.

## Changelog

> Insert changelog entry or delete this section.