Fix using the stop_timeout argument in DockerRunLauncher.container_kwargs (#25423)
## Summary & Motivation
container_kwargs.stop_timeout should work out of the box, but does not
due to a longstanding docker-py bug. Move it into the terminate() method
instead

## How I Tested These Changes
New test case

## Changelog
[dagster-docker] `container_kwargs.stop_timeout` can now be set when
using the `DockerRunLauncher` or `docker_executor` to configure the
amount of time that Docker will wait when terminating a run for it to
clean up before forcibly stopping it with a SIGKILL signal.