[1.9] Update default value of allow_backfills to True (#25204)
## Summary & Motivation

This updates the default value of this setting. To be merged in 1.9.

## How I Tested These Changes

## Changelog

By default, `AutomationConditionSensorDefinitions` will now emit backfills to handle cases where more than one partition of an asset is requested on a given tick. This allows that asset's BackfillPolicy to be respected. This feature can be disabled by setting `allow_backfills` to `False`.