[1.9] Promote hidden params to top-level args in AutomationConditionSensorDefinition (#25206)
## Summary & Motivation

As title -- these should be documented and visible to users.

Open to suggestions on `use_user_code_server` as a name.

## How I Tested These Changes

## Changelog

Added a new `use_user_code_server` paramter to `AutomationConditionSensorDefinition`. If set, the sensor will be evaluated in the user code server (as traditional sensors are), allowing custom `AutomationCondition` subclasses to be evaluated. To learn more, view the docs [TODO].

Added a new `default_condition` parameter to `AutomationConditionSensorDefinition`. If set, this condition will be used for any assets or asset checks within the selection that do not have an automation condition defined. Requires `use_user_code_server` to be set.