[ui] Re-order logic in Materialize button to avoid pipeline resource check (#25417)
## Summary & Motivation

In this thread it was identified that there is a query waterfall in the
Materialize button:
https://dagsterlabs.slack.com/archives/C070G3UNNLW/p1724874261476729

These queries can be surprisingly expensive, taking 5-6s in some DD
traces. There are three primary queries we make:

1) Loading the assets in the selection to get their dependencyKeys,
backfill policies, partition definitions, asset checks and required
resources.

2) Loading the assets upstream of the selection (which requires the
dependencyKeys from (1)) with their last materialization to see if any
upstream assets have never been materialized, in which case this
materialization will likely fail. (This warning might not be worth the
additional time, or the complexity of refactoring it to pull cached data
/ be faster)

3) Loading the pipeline that will be used to materialize the assets to
determine if the resources required by the assets have required config,
which means the launchpad must be displayed.

**The third one is only necessary sometimes - you have to have to be
launching an un-partitioned asset in a custom job with a resource for it
to be relevant.**. I've made a very careful re-org of the logic in this
third test, allowing to be skipped most of the time.

In the future, it might be nice to actually hit the materialize mutation
and have it error, kicking the user to the launchpad

## How I Tested These Changes

I traced this logic carefully to verify that it produces the same output
as the old logic. We actually have a test for this code as well
(`materialize button > should show the launchpad if an asset or resource
requires config`).


![image](https://github.com/user-attachments/assets/9a266aed-31a0-41a1-800c-e32ac46cec74)

Based on traces like the one above, I expect this to make the overall
flow 10% faster. Not a ton but enough to be worth it I think.

## Changelog

[ui] The Materialize button runs pre-flight checks more efficiently,
resulting in a 10% improvement in time-to-launch.

Co-authored-by: bengotow <bgotow@elementl.com>