[dagster-pipes] feat: Multi-container log streaming (#24313)
## Summary & Motivation
Note: This is as pretty massive PR and I don't necessarily intend to get
it merged in its current form. However I would like to initiate the
discussion on the approach being taken and go from there. cc
@alangenfeld

This PR intends to do two things:
1. Enable streaming of logs from multiple containers in Dagster.
Previously, since the `PipesK8sPodLogsMessageReader.consume_pod_logs`
function doesn't specify a container name to `read_namespaced_pod_log`,
it would error if your pipes pod had sidecars.
2. Address https://github.com/dagster-io/dagster/issues/21331. Now, if
the log stream dies from the container, Dagster will refresh the log
stream and continue to wait for the pod.

We have been using this custom code in our codebase for months without
issue. Unfortunately, the solutions to 1. and 2. are quite intertwined
and were not easy for us to split into two PRs for the Dagster codebase.
Broadly, this PR achieves these goals by doing the following:

- Create a stream of log streams for each container
- Use queues to process each stream in separate worker tasks
- Refresh the log stream when it dies
- Order the logs from each log stream in a Priority Queue, and
deduplicate any logs
- Ensure streams exit gracefully if pod is killed and all tasks are
cleaned up

The change has been gated behind a flag to the pipes `run` command
`enable_multi_container_logs`. This PR as is will not be sufficient as
I'm sure the Dagster team will wish that this flag can be passed in via
config (e.g. Helm) but I am not sure if it is preferable to deploy this
stealthily first with no way to expose it in config. Let me know.

## How I Tested These Changes
- We've ran dozens of jobs using this code
- We have run jobs that exceed the "4 hour limit". In those cases we
notice the streams refresh and logs are deduplicated correctly
- Created extensive tests for this logic

## Changelog

Insert changelog entry or "NOCHANGELOG" here.

- [x] `NEW` _Support multi-container log streaming_
- [x] `BUGFIX` _Fixed pipes K8S op failing after 4hrs_
- [ ] `DOCS` _(added or updated documentation)_

## Notes
The nested context managers and merge stream queues may seem like
over-engineering. We initially attempted this using a more simple
approach that depended more heavily on the threading library to process
the logs as well as wait for the pod termination. However we noticed
this approach was flakey and thought it would be better to wait for pod
termination in the main thread.

---------

Co-authored-by: Matthew Kuzyk <matthew.kuzyk@woven-planet.global>
Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>