Stash, retrieve `StateBackedDefinitionsLoader` reconstruction data in `DefinitionsLoadContext`  (#25436)
## Summary

Currently, `StateBackedDefinitionsLoader` stashes its reconstruction
data in the output `Definitions`. This means that in order for it to be
persisted, that `Definitions` object must make its way to the terminal
user-specified `defs`. In the case that we build APIs which emit
non-Definitions objects as a result of cached computation (e.g. specs),
we lose that reconstruction data.

This PR moves to a paradigm where we store this reconstruction data in
the `DefinitionsLoadContext`, where it's retrieved and attached to the
final `RepositoryDefinition`.

## Test Plan

Existing tests.