[lazy-defs] Eliminate PendingRepositoryDefinition (#24632)
## Summary & Motivation

Completely remove `PendingRepositoryDefinition` from the codebase.

The class is not public (i.e. not a top-level export), though it is
currently returned by `@repository` if at least one
`CacheableAssetsDefinition` is passed. Given that it has no public APIs
and isn't a top-level export, I think it is OK to eliminate it.

The replacement behavior is to resolve `CacheableAssetsDefinition`
immediately inside the `@repository` decorator. This is made possible by
bringing in the global load context via `DefinitionLoadContext.get()`.
If we are in reconstruction mode, then we can access previously
generated `cacheable_asset_data` off the `DefinitionLoadContext`. If we
are in initialization mode, then we compute the cacheable asset data and
stash it on the `RepositoryLoadData` that we are already generating
inside the `@repository` decorator.

The biggest effect of this change is that cacheable asset data
generation routines are now called immediately when
`CacheableAssetsDefinition` is passed into `@repository`, which
basically means at import time. The previous behavior was to only call
these routines once `repository_def_from_target` was called on the
`PendingRepositoryDefinition` created at import time.

I am not sure whether this change has a practical impact on users, since
the framework calls `repository_def_from_target` shortly after import
whenever a code location is loaded. It did, however, have a significant
impact on our tests. I needed to wrap several inlined "pending repo"
definitions in `@lazy_definitions` to avoid these routines being
triggered when a test file was loaded. This was important because (a)
some tests require loading the repo multiple times; (b) some of the
cacheable data generation routines used in tests require a defined
`DagsterInstance` which doesn't exist at the top level of a test file. I
also renamed all the "pending repo" instances to "cacheable assets repo
(or defs)" since `PendingRepositoryDefinition` is now gone.

## How I Tested These Changes

Existing test suite (with some modifications).