filter.exclude_subruns uses backfill_id column if present (#25449)
## Summary & Motivation
If the data migration to populate the `RunsTable.backfill_id` column has
been run, we can use the column to power the `exclude_subruns` filter

## How I Tested These Changes
updated backcompat tests to show that querying runs with the
`exclude_subruns` filter works before and after the migration.

Also ran some perf tests locally against a DB with ~9 million runs and
observed that the Runs Feed loads in about a second

## Changelog

Added a new column to the Runs table to improve performance of the Runs
page. To take advantage of these performance improvements, run `dagster
instance migrate`. This migration involves a schema migration to add the
new column, and a data migration to populate the new column for
historical runs.