[bk] Update individual package publish pipeline to handle rcs (#25352)
## Summary

Updates the logic of our single-package publish BK pipeline to:
1) Allow selecting any package, not just those with fixed versions in
the codebase
2) If a package is selected that doesn't have a fixed version in the
codebase [meaning, it is a package released weekly with typical dagster
release version numbers] mandate that the release version is a rc, so it
doesn't interrupt typical user installs

The intent is that we can use this pipeline to (1) freeform release
packages we don't regularly publish, e.g. `dagster-airlift` and (2)
allow us to easily release release candidates for packages we do
regularly publish, e.g. `dagster-looker`

## Test Plan

Successful release: 
https://buildkite.com/dagster/publish-single-python-package/builds/56#

Flagging non-rc version for traditional package:
https://buildkite.com/dagster/publish-single-python-package/builds/54