Update dagster-helm to use pydantic 2 (#25507)
Internal companion PR: https://github.com/dagster-io/internal/pull/12359

## Summary & Motivation

Update `dagster-helm` to use Pydantic 2+ instead of Pydantic 1. This
unblocks updating `dagster` to use exclusively 2+ (Pydantic 1.x is EOL).

Because the entire purpose of `dagster-helm` is to define our helm
schema with pydantic, and Pydantic 1 -> 2 has many breaking changes,
there are a lot of updates here:

- Fields of form: `field: Optional[SomeType]` were updated to `field:
Optional[SomeType] = None`. The `None` default is automatically applied
in v1 but not v2.
- `Config` classes in our models were removed in favor of
`model_config=ConfigDict(...)` or setting config directly in the class
args. This is the preferred API in v2.
- Renamed `schema_extra` config specifications to `json_schema_extra`
- Kubernetes models with no defined fields but a `json_schema_extra`
pointing to an external schema now have `extra=allow` on them. This
allows fields to actually be set on these objects-- this was possible in
v1 because `BaseModel.construct()` ignored the `extra` setting and
always set extra fields, but that behavior changed in v2.
- Tests that constructed an object for a schema with nested models
previously sometimes just used a dict instead of a model object on the
test object. This stopped working and now we make sure the test object
contains all the nested model objects.
- Few other tweaks for niche API breakages.

## How I Tested These Changes

Existing test suite.