Fix "Large launchpad configs returned by Test Sensor feature causes URL to be too long and page to be unloadable" (#25577)
## Summary & Motivation
As titled, fixes the above issue.

So the original implementation redirects to the `/setup` path which all
it does is save the configuration to localStorage and then immediately
redirects to the `/playground` route. This solution cuts out the
middleman and just directly saves the config to localStorage and then
opens `/playground` directly. This avoids the need to pass a large
config via the URL.

## How I Tested These Changes

Used the test sensor feature and opened one of the config's it returned.