[core] modify context type signatures and usage for making AssetExecutionContext a standalone class (#25541)
## Summary & Motivation

Now that all of the libraries have been updated to support
`AssetExecutionContext`, we can make `AssetExecutionContext` not a
subclass of `OpExecutionContext`

`AssetExecutionContext` is marked to be made a standalone class (ie not
a subclass of `OpExecutionContext`) in 1.9. These are the changes in the
core dagster module that are required to make that happen. It's just
removing the metaclass that enables the `isinstance` check to raise a
deprecation warning, and removing the actual inheritance.

I also delete the tests that ensure `OpExecutionContext` and
`AssetExecutionContext` implement the same methods since we don't
need/want to enforce that now that they are separate classes.

## How I Tested These Changes
existing tests

## Changelog

`AssetExecutionContext` is no longer a subclass of `OpExecutionContext`.
At this release, `AssetExecutionContext` and `OpExecutionContext`
implement the same methods, but in the future, the methods implemented
by each class may diverge. If you have written helper functions with
`OpExecutionContext` type annotations, they may need to be updated to
include `AssetExecutionContext` depending on your usage. Explicit calls
to `isinstance(context, OpExecutionContext)` will now fail if `context`
is an `AssetExecutionContext`.