[snowflake] Support ignoring missing tables when using `fetch_last_updated_timestamps` (#25488)
## Summary & Motivation
This PR adds an optional argument `ignore_missing_tables: Optional[bool]
= False` to `dagster_snowflake.fetch_last_updated_timestamps`. When
true, this argument will not raise an error when a table is not found in
Snowflake. Instead, it will just continue and exclude the table from the
resulting freshness map.

This is useful when you deploy a dagster project to multiple
environments, where your development environment may not have all tables
available. We use this function to generate observation events for
source tables, and this allows us to generate observations for the
tables that do exist, but ignore tables that don't.

Linked issue: https://github.com/dagster-io/dagster/issues/24245

## How I Tested These Changes

Unit test is included.

## Changelog

> `dagster_snowflake.fetch_last_updated_timestamps` now supports
ignoring tables not found in Snowflake instead of raising an error