[ui] Handle failure in code location permalink (#25625)
## Summary & Motivation

When a code location is in an error state, link to its permalink so that
metadata is available.

The current app behavior is to show an empty state when attempting to
load an unloadable code location. This change replaces the empty state
with the code location overview page.

If the code location has no repositories available (because it's in an
error state), disable the "Definitions" tab.


https://github.com/user-attachments/assets/714141b6-ac88-485e-bad2-7764482f0194


## How I Tested These Changes

Load the app locally, cause a Python error in `toys`, reload the code
location. Verify that the code location goes into an errored state but
is linked from the Code locations page.

Click the link, verify that I arrive at the code location permalink,
with metadata and the code location error available.