[dagster-dlift] implement translator (#25482)
## Summary & Motivation
Scaffolds out pattern for translator and spec construction. Follows almost exactly the pattern laid out by BI integrations, except for the fact that the project in this case is not a resource (I find that pattern to be somewhat unnecessary here).
## How I Tested These Changes
New tests for asset retrieval and project construction. Added a "jaffle shop" fake built on top of the previous unit testing utilities which we instead base all the tests on.