[ui] Fix backfill coordinator logs not transitioning from loading => empty (#25714)
## Summary & Motivation

There was a report during dogfooding that the Coordinator logs tab does
not transition from loading to empty for old backfills like this one:
https://dogfood-test-1.dogfood.dagster.cloud/prod/runs/b/zjexmmzx?tab=logs

This UI actually implements both loading and empty states, but the empty
state wasn't reached because `useCursorAccumulatedQuery` uses an
`AccumulatingDataFetcher`, and the fetcher's `.fetch()` method never
called `onData` if the first page of results was empty. I think it
should so that consumers have a way to transition from loading > empty.

## How I Tested These Changes

I tested these changes on the coordintor logs page viewing the backfill
linked above. At the moment, this is the only place
`AccumulatingDataFetcher` or `useCursorAccumulatedQuery` are used. It
looks like they still need tests, but I'll add them in a follow-up.
Before we write tests for the current implementation, I think it'd make
sense to try to use it in at least one more place (seems like are other
log views that it would work for??) and see if other changes are
necessary.

## Changelog

[ui] The backfill coordinator logs tab no longer sits in a loading state
when no logs are available to display.

Newly-visible empty state:
<img width="1840" alt="image"
src="https://github.com/user-attachments/assets/85c733a1-0b13-4ef4-9e04-4a4aea049b56">

Co-authored-by: bengotow <bgotow@elementl.com>